worker/requirements.txt
sqlalchemy==2.0.37
psycopg[binary]==3.2.4
PyYAML==6.0.2